Move SslContext creation into a separate builder.
